requirements.txt
Flask==3.1.1
requests==2.32.3
gunicorn==23.0.0
Jinja2==3.1.6
python-dotenv==1.1.0